Se agrego cropper en pantallas editar mascota
